requirements.txt
# Pillow
Django
psycopg2
envs
django-extensions
ipython
jupyter
djangorestframework
flake8
